refactor Commands
implements #1
